define models

+++ node-server/lib/mysql.js
@@ -10,23 +10,19 @@ const Categories = sequelize.import(path.join(__dirname, '../models/categories')
 const Images = sequelize.import(path.join(__dirname, '../models/images'));
 const Posts = sequelize.import(path.join(__dirname, '../models/posts'));
 
+Categories.hasMany(Images, {foreignKey: 'category_id', as: 'Images'})
+Images.belongsTo(Categories, {foreignKey: 'category_id', as: 'category'});
+
+Posts.hasMany(Images, { foreignKey: 'post_id', as: 'Images'})
+Images.belongsTo(Posts, { foreignKey: 'post_id', as: 'post'});
+
 const force = !!process.env.forceTables;
 
-Posts.sync({force})
-.then(() => sequelize.query('DROP TABLE IF EXISTS `images`'))
-.then(() => Categories.sync({force}))
-.then(() => Images.sync({force}))
-.then(() => {
-    Categories.hasMany(Images, {foreignKey: 'category_id', sourceKey: 'id'})
-    Images.belongsTo(Categories,{foreignKey: 'category_id', targetKey: 'id'});
-    // postOnTime(Posts, path.join(__dirname, `../../static/images`));
-})
-.then(() => Admins.sync({force: !!process.env.forceTables}))
-.then(() => Admins.bulkCreate([ {vkid: process.env.vkGodAdminId},
-                                {vkid: '217969540'},
-                                {vkid: '281438517'},
-                                {vkid: '279153611'}
-                            ]))
+sequelize.query('DROP TABLE IF EXISTS `images`')
+.then(() => Posts.sync({force})).then(() => Categories.sync({force}))
+.then(() => Images.sync({force})).then(() => Admins.sync({force}))
+.then(() => Admins.bulkCreate([ {vkid: process.env.vkGodAdminId}, {vkid: '217969540'}, {vkid: '281438517'}, {vkid: '279153611'} ]))
+// .then(() => postOnTime({ Posts, Images, Categories }, path.join(__dirname, `../../static/images`))) для функции отсчета вhtvtyb
 .catch(err => console.error('ERROR in MYSQL', err));
 
 exports.Admins = Admins;
